+++ react/tic-tac-toe/src/app/screens/Game/components/Board/index.js
@@ -5,26 +5,6 @@ import Square from '../Square';
 import styles from './styles.module.scss';
 
 class Board extends Component {
-  calculateWinner = (squares) => {
-    const lines = [
-      [0, 1, 2],
-      [3, 4, 5],
-      [6, 7, 8],
-      [0, 3, 6],
-      [1, 4, 7],
-      [2, 5, 8],
-      [0, 4, 8],
-      [2, 4, 6]
-    ];
-    for (let i = 0; i < lines.length; i++) {
-      const [a, b, c] = lines[i];
-      if (squares[a] && squares[a] === squares[b] && squares[a] === squares[c]) {
-        return squares[a];
-      }
-    }
-    return null;
-  }
-
   renderSquare(i) {
     return (
       <Square
@@ -37,7 +17,6 @@ class Board extends Component {
   render() {
     return (
       <div>
-        {/* <div className={styles.status}>{status}</div> */}
         <div className={styles.boardRow}>
           {this.renderSquare(0)}
           {this.renderSquare(1)}
