+++ react/tic-tac-toe/src/app/screens/Game/index.js
@@ -2,6 +2,7 @@ import React, { Component } from 'react';
 
 import styles from './styles.module.scss';
 import Board from './components/Board';
+import { calculateWinner } from './utils';
 
 class Game extends Component {
   state = {
@@ -12,31 +13,11 @@ class Game extends Component {
     xIsNext: true
   };
 
-  calculateWinner = (squares) => {
-    const lines = [
-      [0, 1, 2],
-      [3, 4, 5],
-      [6, 7, 8],
-      [0, 3, 6],
-      [1, 4, 7],
-      [2, 5, 8],
-      [0, 4, 8],
-      [2, 4, 6]
-    ];
-    for (let i = 0; i < lines.length; i++) {
-      const [a, b, c] = lines[i];
-      if (squares[a] && squares[a] === squares[b] && squares[a] === squares[c]) {
-        return squares[a];
-      }
-    }
-    return null;
-  }
-
-  handleClick = (i) => {
+  handleClick = i => {
     const history = this.state.history.slice(0, this.state.stepNumber + 1);
     const current = history[history.length - 1];
     const squares = current.squares.slice();
-    if (this.calculateWinner(squares) || squares[i]) {
+    if (calculateWinner(squares) || squares[i]) {
       return;
     }
     squares[i] = this.state.xIsNext ? 'X' : 'O';
@@ -49,18 +30,16 @@ class Game extends Component {
     });
   }
 
-  jumpTo = (step) => {
+  jumpTo = step =>
     this.setState({
       stepNumber: step,
       xIsNext: step % 2 === 0
     });
-  }
-
 
   render() {
     const history = this.state.history;
     const current = history[this.state.stepNumber];
-    const winner = this.calculateWinner(current.squares);
+    const winner = calculateWinner(current.squares);
 
     const moves = history.map((step, move) => {
       const desc = move
@@ -88,7 +67,7 @@ class Game extends Component {
           />
         </div>
         <div className={styles.gameInfo}>
-          <div>{status}</div>
+          {status}
           <ol>{moves}</ol>
         </div>
       </div>
